NetApp-main/app/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.2
pymongo==4.8.0
boto3==1.34.162
minio==7.2.7
azure-storage-blob==12.23.1
google-cloud-storage==2.18.2
requests==2.32.3
kafka-python==2.0.2
python-dotenv==1.0.1
